+++ lib/schleuder/mail/message.rb
@@ -336,6 +336,15 @@ module Mail
     end
 
     def add_list_headers(list)
+      if list.include_autocrypt_header
+        # Inject whitespaces, to let Mail break the string at these points
+        # leading to correct wrapping.
+        keydata = list.export_key_minimal_base64.gsub(/(.{78})/, '\1 ')
+        
+        # Inject another whitespace to ensure visual alignment.
+        self['Autocrypt'] = "addr=#{list.email}; prefer-encrypt=mutual; keydata= #{keydata}"
+      end
+
       if list.include_list_headers
         self['List-Id'] = "<#{list.email.gsub('@', '.')}>"
         self['List-Owner'] = "<mailto:#{list.owner_address}> (Use list's public key)"